Remove side-effectful registration of A/B tests
