+++ packages/lesswrong/lib/abTests.ts
@@ -1,4 +1,4 @@
-import { ABTest } from './abTestImpl';
+import { ABTest } from "./abTestClass";
 
 /*
  * A/B tests, both active and finished. If an A/B test is active--that is,
@@ -28,7 +28,6 @@ import { ABTest } from './abTestImpl';
  *     }
  *   });
  */
-
 
 export const welcomeBoxABTest = new ABTest({
   name: "welcomeBoxABTest",
@@ -63,3 +62,5 @@ export const twoLineEventsSidebarABTest = new ABTest({
     },
   },
 });
+
+export const allABTests = [welcomeBoxABTest, twoLineEventsSidebarABTest];
